+++ src/sys/mem/mod.rs
@@ -32,6 +32,11 @@ mod pin;
 #[cfg_attr(feature = "nightly_doc", doc(cfg(feature = "unsafe_layout")))]
 mod pod;
 
+#[cfg(feature = "std")]
+#[cfg(all(not(feature = "safe_mem"), feature = "unsafe_layout"))]
+#[cfg_attr(feature = "nightly_doc", doc(cfg(all(feature = "std", feature = "unsafe_layout"))))]
+mod guard; // Current, CurrrentGuard
+
 pub mod cell;
 
 crate::items! { // structural access: _mods, _pub_mods, _hidden, _all, _always
@@ -49,6 +54,9 @@ crate::items! { // structural access: _mods, _pub_mods, _hidden, _all, _always
         pub use super::pin::Pinned;
         #[cfg(all(not(feature = "safe_mem"), feature = "unsafe_layout"))]
         pub use super::pod::MemPod;
+        #[cfg(feature = "std")]
+        #[cfg(all(not(feature = "safe_mem"), feature = "unsafe_layout"))]
+        pub use super::guard::{Current, CurrentGuard};
         // WIPZONE
         // pub use super::borrow::*;
     }